+++ src/api/consumer/lib/test-inline.js
@@ -3,9 +3,16 @@ import { loadConsumer } from '../../../consumer';
 import InlineId from '../../../consumer/bit-inline-id';
 import Bit from '../../../consumer/component';
 
-export default function test(id: string): Promise<Bit> {
+export function testInline(id: string): Promise<Bit> {
   return loadConsumer()
     .then(consumer =>
       consumer.runComponentSpecs(InlineId.parse(id))
+    );
+}
+
+export function testInlineAll(): Promise<Bit> {
+  return loadConsumer()
+    .then(consumer =>
+      consumer.runAllInlineSpecs()
     );
 }